+++ example/play.ts
@@ -5,9 +5,9 @@ console.log(
   generateDeclaration({
     name: 'piral-sample',
     root: resolve(__dirname, '..', 'tests', 'assets'),
-    files: ['merge1.ts', 'merge1-one.ts', 'merge1-two.ts'],
-    types: ['merge1.ts'],
+    files: ['defaults4.ts'],
+    types: ['defaults4.ts'],
     apis: [],
-    imports: [],
+    imports: ['react'],
   }),
 );