+++ CHANGELOG.md
@@ -6,6 +6,15 @@
 - Added import equals declaration
 - Added support for ref comments (#14)
 
+## 0.5.5
+
+- Ignore floating import nodes
+
+## 0.5.4
+
+- Fixed problem with default export in ambient modules
+- Improved support for default (class) export
+
 ## 0.5.3
 
 - Fixed invalid file references